+++ crates/xtask/src/cmd/compare/mod.rs
@@ -4,10 +4,12 @@ mod all;
 mod diagrams;
 mod layout;
 mod paths;
+mod root;
 mod xml;
 
 pub(crate) use all::*;
 pub(crate) use diagrams::*;
 pub(crate) use layout::*;
 pub(crate) use paths::*;
+pub(crate) use root::*;
 pub(crate) use xml::*;